Update main tutorial and small changes
* Move rmse_3d to utils as it is general
* Extra print statements for verbose_weights
